Update the database to MySQL

+++ db/migrate/20140817023351_remove_product_attributes.refinery_ironman.rb
@@ -2,7 +2,9 @@
 class RemoveProductAttributes < ActiveRecord::Migration
 
   def up
-    remove_column :refinery_ironman_products, :product_attribs
+    if column_exists? :refinery_ironman_products, :product_attribs
+      remove_column :refinery_ironman_products, :product_attribs
+    end
   end
 
   def down
